Modify active indicators
Active indicators are now all circles, however without a typing indicator. This may be added back in the future, but the headache it was causing trying to move the default one wasn't worth it.

Also, when you hover over an avatar, the "VIEW PROFILE" overlay no longer dims the image.